DBZ-4353 Add online ddl case in VitessConnectorIT
